api/pyproject.toml
[project]
name = "oscillink-latticedb"
version = "0.1.0"
description = "Local lattice-of-lattices RAG database scaffold with receipts and verification"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.7.4",
  "pydantic-settings==2.4.0",
  "PyJWT==2.10.1",
  "numpy==1.26.4",
  "pyarrow==16.1.0",
  "pandas==2.2.3",
  "scikit-learn==1.4.2",
  "scipy==1.12.0",
  "orjson==3.10.7",
  "typing-extensions==4.12.2",
  "prometheus-client==0.20.0",
  "PyYAML==6.0.2"
]
[project.optional-dependencies]
dev = [
  "pytest==8.2.2",
  "pytest-cov==5.0.0",
  "httpx==0.27.2",
  "requests==2.32.3",
  "ruff==0.5.7",
  "pre-commit==3.7.1",
  "pooch==1.8.2",
  "pdfminer.six==20231228",
  "python-docx==1.1.2",
  "faiss-cpu>=1.7.4",
  "psutil>=5.9.0"
]
embeddings = [
  "transformers==4.44.2",
  "torch>=2.0.0",
  "sentencepiece>=0.1.99",
  "accelerate>=0.33.0"
]
[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]